streamlit/requirements.txt
altair
pandas
streamlit
PyJWT[crypto]==2.9.0
requests==2.32.3